+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Arch
     - 'clean cache' operation was not working in some scenarios
     - upgrading progress when conflicting files are detected
+    - not detecting some installed 'not-signed' repository packages
     
 - upgrade: crashing for scenarios when there are packages that cannot upgrade to be displayed on the summary window
 - settings: crashing when an empty Qt style is defined [#104](https://github.com/vinifmor/bauh/issues/104)